+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
   - Intial start on header macro in jinja templates to make drilldown headings more repeatable
   - initial work on using panopticas for file type identification
   - added some static methods for generating and parsing repo_id in kospex_git
+  - Implemented command on kospex CLI for feature request [kospex version command](https://github.com/kospex/kospex/issues/13)
 
 ### Changed
   - Fixed how percentages and circles were created in summary view